Update POST /api/users to meet MVP requirements
Originally I was only returning the id for the newly-created user. This
commit adds the rest of the user's properties (name, bio, created_at, etc.)
to the user object before returning. This is done by calling `.findById()`
with the new id.

Also in this commit is a fix for a bug in this same route. The bug would
occur if either name or bio properties were missing, which would correctly
respond to user with the error, however, the `.insert()` method would still
be called later in this route handler. To fix this, I added a simple `return`
statement before `res.status(404)...` to explicitly end the handler execution.